+++ lib/sidekiq_expiring_jobs/middleware.rb
@@ -5,8 +5,11 @@ module SidekiqExpiringJobs
     class Client
       def call(_worker_class, job, _queue, _redis_pool)
         if (expires_in = job.delete("expires_in"))
+          expires_in = expires_in.to_f
+          raise ArgumentError, ":expires_in must be a relative time, not absolute time" if expires_in > 1_000_000_000
+
           at = job["at"] || job["created_at"]
-          job["expires_at"] = at + expires_in.to_f
+          job["expires_at"] = at + expires_in
         end
         yield
       end