+++ test/jobs.rb
@@ -19,6 +19,10 @@ class ExpiringJob100 < BaseJob
   sidekiq_options expires_in: 100
 end
 
+class AbsTimeExpiringJob < BaseJob
+  sidekiq_options expires_in: (Time.now.to_i + 10) # aka 10.seconds.from_now in Rails
+end
+
 class TerminateJob
   include Sidekiq::Job