+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+- Raise when `:expires_in` is an absolute time
+
 ## 0.1.0 (2023-04-12)
 
 - First release